requirements.txt
pdf2image==1.16.3
Pillow==10.0.0
poppler-utils==0.1.0
PyPDF2==3.0.1
pytesseract==0.3.10